refactor(numeric): share probability validation
